fix(host): change iplong from int to int64.

+++ shodan/host.go
@@ -64,7 +64,7 @@ type HostData struct {
 	Hostnames    []string               `json:"hostnames"`
 	Version      HostVersion            `json:"version"`
 	Title        string                 `json:"title"`
-	IPLong       int                    `json:"ip"`
+	IPLong       int64                  `json:"ip"`
 	IP           string                 `json:"ip_str"`
 	OS           string                 `json:"os"`
 	Organization string                 `json:"org"`
@@ -89,7 +89,7 @@ type HostData struct {
 type Host struct {
 	OS              string      `json:"os"`
 	Ports           []int       `json:"ports"`
-	IPLong          int         `json:"ip"`
+	IPLong          int64       `json:"ip"`
 	IP              string      `json:"ip_str"`
 	ISP             string      `json:"isp"`
 	Hostnames       []string    `json:"hostnames"`
